Adds milestone edit modal
